+++ src/soulstruct/base/animations/tae/__init__.py
@@ -4,8 +4,13 @@ __all__ = [
     "TAEEvent",
     "TAEEventGroup",
     "TAEEventType",
+    "RemoTAE",
+    "RemoTAEAnimation",
+    "RemoTAEEvent",
+    "RemoTAEEventGroup",
 ]
 
 from .core import *
 from .enums import *
+from .remo_tae import *
 # `TAEEventData` subtypes not imported here.